terminado, faltan los bonus
