+++ starter-code/views/main/index.hbs
@@ -1,0 +1,4 @@
+<h1>Main Page</h1>
+<a href="/">Go back to home page</a>
+<img src="/images/funnyCat.jpg" alt="funnyCat">
+<a href="/private">Go to private page</a>
