+++ starter-code/views/private/index.hbs
@@ -1,0 +1,4 @@
+<h1>Private Page</h1>
+<p>You're in your private page</p>
+<a href="/">Go back to home page</a>
+<iframe src="https://giphy.com/embed/26FPCXdkvDbKBbgOI" width="480" height="419" frameBorder="0" class="giphy-embed" allowFullScreen></iframe><p><a href="https://giphy.com/gifs/justin-cat-trippy-psychedelic-26FPCXdkvDbKBbgOI">via GIPHY</a></p>
